Slider in app.js
